requirements.txt
certifi
charset-normalizer
idna
python-dotenv
pytz
requests
urllib3
